move parse example serialization tests to kernel tests
