+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -302,22 +302,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "parse_example_dataset_serialization_test",
-    size = "medium",
-    srcs = ["parse_example_dataset_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/experimental/kernel_tests:reader_dataset_ops_test_base",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-    ],
-)
-
-tf_py_test(
     name = "prefetch_dataset_serialization_test",
     size = "small",
     srcs = ["prefetch_dataset_serialization_test.py"],
